backend/ai/gemini_helper.py
import google.generativeai as genai

def generate_content_with_gemini(api_key, model, contents):
    """
    Initializes the Gemini client and generates content using the specified model.
    
    :param api_key: Your Gemini API key.
    :param model: The model to use (e.g., "gemini-2.0-flash").
    :param contents: A list where:
                     - contents[0] is the prompt string,
                     - Optionally, contents[1] is a PIL.Image object if needed.
    :return: The response object from Gemini.
    """
    # Configure the client with your API key.
    genai.configure(api_key=api_key)
    
    # In this example, we'll assume that the image is used to enrich the prompt.
    # Gemini's API may require images to be pre-processed or encoded; adjust as needed.
    prompt = contents[0]
    # For now, we'll ignore the image object in contents[1] if provided.
    
    # Call Gemini's generate_text method (adjust parameters as needed for your model).
    response = genai.generate_text(
        model=model,
        prompt=prompt,
    )
    return response
