backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django==5.1.6",
    "djangorestframework==3.15.2",
    "django-cors-headers==4.6.0",
    "django-environ==0.12.0",
    "dj-database-url==2.3.0",
    "google-generativeai==0.8.4",
    "pillow==11.1.0",
    "requests==2.32.3"
]
